% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{gof}
\alias{gof}
\title{Calculate Goodness-of-Fit Statistics}
\usage{
gof(observed, predicted, type = "spec")
}
\arguments{
\item{observed}{A numeric vector of the true, observed values.}

\item{predicted}{A numeric vector of the model's predicted values.}

\item{type}{The type of output required. Use `"spec"` for detailed
spectrometric model evaluation or `"DSM"` for a simpler set.}
}
\value{
A data frame containing the calculated performance metrics.
}
\description{
Computes a comprehensive set of performance metrics for evaluating a model's
predictions against observed values.
}
\examples{
\dontrun{
# Calculate performance metrics
performance_metrics <- gof(observed = validation_data$tc.perc, predicted = predictions)
print(performance_metrics)
}
}
