% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_isc_spectra}
\alias{read_isc_spectra}
\title{Read and Combine ISC Spectra Files}
\usage{
read_isc_spectra(
  path,
  pattern = "_r.csv",
  id_regex = "(?<=_)(\\\\d+)(?=\\\\_)",
  id_prefix = "isc-",
  skip_lines = 28
)
}
\arguments{
\item{path}{The full file path to the directory containing the raw spectra files.}

\item{pattern}{A character string containing a regular expression to identify
the correct files. Defaults to "_r.csv".}

\item{id_regex}{A regular expression used to extract the sample ID from the
filename. The default pattern `(?<=_)(\\d+)(?=\\_)` looks for digits
enclosed by underscores.}

\item{id_prefix}{A character string to prepend to the extracted numeric ID.
Defaults to "isc-".}

\item{skip_lines}{The number of header lines to skip at the top of each
CSV file. Defaults to 28.}
}
\value{
A tidy tibble (a modern data frame) with three columns: `id`,
  `wavelength`, and `reflectance`.
}
\description{
This function reads all raw CSV files from an ISC spectrometer located in a
specified directory. It combines them into a single, tidy data frame and
extracts a sample ID from each filename.
}
\examples{
\dontrun{
# Define the path to the folder containing your raw spectral files
spec_dir <- "C:/BSSS_project_lowcost_nir/database/raw_data/spectra_usa_ge/isc/"

# Run the function to import and clean the data
my_spectra <- read_isc_spectra(path = spec_dir)

# Print the first few rows of the resulting data frame
print(head(my_spectra))
}
}
