% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{resample_spectra}
\alias{resample_spectra}
\title{Resample Spectra to a New Wavelength Resolution}
\usage{
resample_spectra(spectra, current_wavelengths, new_wavelengths)
}
\arguments{
\item{spectra}{A numeric matrix or data frame of spectra.}

\item{current_wavelengths}{A numeric vector of the original wavelengths.}

\item{new_wavelengths}{A numeric vector of the desired new wavelengths.}
}
\value{
A matrix with the resampled spectra. Column names are set to the
  new wavelengths.
}
\description{
A wrapper around `prospectr::resample` to standardize spectral resolution
using spline interpolation.
}
\examples{
\dontrun{
# Resample to a 5 nm interval between 950 and 1650 nm
new_wavs <- seq(950, 1650, by = 5)
resampled_spectra <- resample_spectra(my_spectra, old_wavs, new_wavs)
}
}
