% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{smooth_sgolay}
\alias{smooth_sgolay}
\title{Apply Savitzky-Golay Smoothing or Derivatives}
\usage{
smooth_sgolay(
  spectra,
  derivative_order = 0,
  polynomial_order = 2,
  window_size = 11
)
}
\arguments{
\item{spectra}{A numeric matrix or data frame of spectra.}

\item{derivative_order}{The order of the derivative to compute (m). Use 0 for
smoothing only. Defaults to 0.}

\item{polynomial_order}{The order of the polynomial to fit (p). Defaults to 2.}

\item{window_size}{The size of the moving window (w). Must be an odd number.
Defaults to 11.}
}
\value{
A matrix with the transformed spectra.
}
\description{
A wrapper around `prospectr::savitzkyGolay` to smooth spectra or compute
derivatives.
}
\examples{
\dontrun{
# Smooth the spectra
smoothed_spectra <- smooth_sgolay(my_spectra, window_size = 11)

# Calculate the first derivative
first_derivative <- smooth_sgolay(my_spectra, derivative_order = 1)
}
}
