% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{train_soil_model}
\alias{train_soil_model}
\title{Train a Predictive Soil Model}
\usage{
train_soil_model(spectra, response, method = "cubist", train_control = NULL)
}
\arguments{
\item{spectra}{A numeric matrix or data frame of spectra (samples as rows).}

\item{response}{A numeric vector of the soil property to be predicted. Must
have the same length as the number of rows in `spectra`.}

\item{method}{The modeling method to be used by `caret`. Defaults to "cubist".}

\item{train_control}{A `trainControl` object from the `caret` package. A
default is provided if none is specified.}
}
\value{
A trained model object of class `train`.
}
\description{
A wrapper around `caret::train` to build a model for predicting a soil
property from spectral data.
}
\examples{
\dontrun{
# Train a Cubist model using default repeated cross-validation
soil_model <- train_soil_model(cal_spectra, cal_soil_data$tc.perc)

# Predict new values
predictions <- predict(soil_model, val_spectra)
}
}
