% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{transform_snv}
\alias{transform_snv}
\title{Apply Standard Normal Variate (SNV) Transformation}
\usage{
transform_snv(spectra)
}
\arguments{
\item{spectra}{A numeric matrix or data frame of spectra.}
}
\value{
A matrix with the SNV-transformed spectra.
}
\description{
A wrapper around `prospectr::standardNormalVariate` to correct for
multiplicative scatter effects.
}
\examples{
\dontrun{
# Apply SNV after smoothing
snv_spectra <- my_spectra \%>\%
  smooth_sgolay() \%>\%
  transform_snv()
}
}
